#570 - Extracted BaseSlice, MainSlice and AllRepositoriesSlice
